+++ src/components/Shortcut/index.ts
@@ -1,0 +1,6 @@
+export * from "./ShortcutContext";
+export * from "./ShortcutProvider";
+export * from "./ShortcutRecorder";
+export * from "./types";
+export * from "./useShortcut";
+export * from "./utils";